feat: add authentication and user management features

+++ src/controllers/URL.js
@@ -1,9 +1,11 @@
 const {StatusCodes} = require('http-status-codes');
-const Url = require('../models/URL');
+const Url = require('../models/Url');
 const generateCode = require('../../utils');
 
 const createShortURL = async (req, res) => {
  const payload = req.body;
+ const {userId} = req.user;
+ payload.ownerId = userId;
  console.log(payload);
  if (!payload.longURL) {
   return res.status(StatusCodes.BAD_REQUEST).json({message: "Long URL is required"});
@@ -14,7 +16,8 @@ const createShortURL = async (req, res) => {
 }
 
 const getShortURL = async (req,res) => {
- const resp = await Url.find({});
+ const userId = req.user.userId;
+ const resp = await Url.find({ownerId: userId});
  res.status(StatusCodes.OK).json({data: resp, count: resp.length});
 }
 
@@ -32,8 +35,9 @@ const redirectToLongURL = async (req,res) => {
 }
 
 const getSingleShortURL = async (req,res) => {
+ const userId = req.user.userId;
  const {id: shortURLId} = req.params;
- const resp = await Url.findOne({code: shortURLId});
+ const resp = await Url.findOne({code: shortURLId, ownerId: userId});
  if (!resp) {
   return res.status(StatusCodes.NOT_FOUND).json({message: "Short URL not found"});
  }
@@ -42,7 +46,8 @@ const getSingleShortURL = async (req,res) => {
 
 const deleteShortURL = async (req, res) => {
  const {id: shortURLId} = req.params;
- const resp = await Url.findOneAndDelete({code: shortURLId});
+ const userId = req.user.userId;
+ const resp = await Url.findOneAndDelete({code: shortURLId, ownerId: userId});
  if (!resp) {
   return res.status(StatusCodes.NOT_FOUND).json({message: "Short URL not found"});
  }
